backend/requirements.txt
fastapi
uvicorn
cryptography
pyopenssl
httpx
python-dotenv
python-multipart
sqlalchemy
asyncpg
passlib
psycopg2-binary 
alembic
pydantic-settings
pydantic[email]
cachetools
python-jose[cryptography]
authlib
itsdangerous